pyproject.toml
[project]
name = "alfred-task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic-ai>=0.0.1",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "linear-api>=0.1.0",
    "pyyaml>=6.0.2",
    "anthropic>=0.34.0",
    "tenacity>=8.2.0",
    "ruff>=0.12.10",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "pytest-asyncio>=0.21.0",
]
